refactor: create DependentResources in Controller
