+++ operator-framework-core/src/main/java/io/javaoperatorsdk/operator/processing/Controller.java
@@ -2,6 +2,7 @@ package io.javaoperatorsdk.operator.processing;
 
 import java.util.LinkedList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,9 +19,11 @@ import io.javaoperatorsdk.operator.MissingCRDException;
 import io.javaoperatorsdk.operator.OperatorException;
 import io.javaoperatorsdk.operator.api.config.ConfigurationServiceProvider;
 import io.javaoperatorsdk.operator.api.config.ControllerConfiguration;
+import io.javaoperatorsdk.operator.api.config.dependent.DependentResourceSpec;
 import io.javaoperatorsdk.operator.api.monitoring.Metrics;
 import io.javaoperatorsdk.operator.api.monitoring.Metrics.ControllerExecution;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
+import io.javaoperatorsdk.operator.api.reconciler.ContextInitializer;
 import io.javaoperatorsdk.operator.api.reconciler.DeleteControl;
 import io.javaoperatorsdk.operator.api.reconciler.EventSourceContext;
 import io.javaoperatorsdk.operator.api.reconciler.EventSourceInitializer;
@@ -28,11 +31,13 @@ import io.javaoperatorsdk.operator.api.reconciler.Ignore;
 import io.javaoperatorsdk.operator.api.reconciler.Reconciler;
 import io.javaoperatorsdk.operator.api.reconciler.UpdateControl;
 import io.javaoperatorsdk.operator.api.reconciler.dependent.DependentResource;
-import io.javaoperatorsdk.operator.processing.dependent.DependentResourceManager;
+import io.javaoperatorsdk.operator.api.reconciler.dependent.DependentResourceConfigurator;
+import io.javaoperatorsdk.operator.api.reconciler.dependent.EventSourceProvider;
+import io.javaoperatorsdk.operator.api.reconciler.dependent.KubernetesClientAware;
 import io.javaoperatorsdk.operator.processing.event.EventSourceManager;
 import io.javaoperatorsdk.operator.processing.event.source.EventSource;
 
-@SuppressWarnings({"unchecked"})
+@SuppressWarnings({"unchecked", "rawtypes"})
 @Ignore
 public class Controller<P extends HasMetadata> implements Reconciler<P>,
     LifecycleAware, EventSourceInitializer<P> {
@@ -43,7 +48,8 @@ public class Controller<P extends HasMetadata> implements Reconciler<P>,
   private final ControllerConfiguration<P> configuration;
   private final KubernetesClient kubernetesClient;
   private final EventSourceManager<P> eventSourceManager;
-  private final DependentResourceManager<P> dependents;
+  private final List<DependentResource> dependents;
+  private final boolean contextInitializer;
 
   public Controller(Reconciler<P> reconciler,
       ControllerConfiguration<P> configuration,
@@ -51,14 +57,43 @@ public class Controller<P extends HasMetadata> implements Reconciler<P>,
     this.reconciler = reconciler;
     this.configuration = configuration;
     this.kubernetesClient = kubernetesClient;
+    contextInitializer = reconciler instanceof ContextInitializer;
 
     eventSourceManager = new EventSourceManager<>(this);
-    dependents = new DependentResourceManager<>(this);
+
+    dependents = configuration.getDependentResources().stream()
+        .map(drs -> createAndConfigureFrom(drs, kubernetesClient))
+        .collect(Collectors.toList());
+  }
+
+  @SuppressWarnings("rawtypes")
+  private DependentResource createAndConfigureFrom(DependentResourceSpec dependentResourceSpec,
+      KubernetesClient client) {
+    DependentResource dependentResource =
+        ConfigurationServiceProvider.instance().createFrom(dependentResourceSpec);
+
+    if (dependentResource instanceof KubernetesClientAware) {
+      ((KubernetesClientAware) dependentResource).setKubernetesClient(client);
+    }
+
+    if (dependentResource instanceof DependentResourceConfigurator) {
+      final var configurator = (DependentResourceConfigurator) dependentResource;
+      dependentResourceSpec.getDependentResourceConfiguration()
+          .ifPresent(configurator::configureWith);
+    }
+    return dependentResource;
+  }
+
+  private void initContextIfNeeded(P resource, Context<P> context) {
+    if (contextInitializer) {
+      ((ContextInitializer<P>) reconciler).initContext(resource, context);
+    }
   }
 
   @Override
   public DeleteControl cleanup(P resource, Context<P> context) {
-    dependents.cleanup(resource, context);
+    initContextIfNeeded(resource, context);
+    dependents.forEach(dependent -> dependent.cleanup(resource, context));
 
     try {
       return metrics().timeControllerExecution(
@@ -90,7 +125,8 @@ public class Controller<P extends HasMetadata> implements Reconciler<P>,
 
   @Override
   public UpdateControl<P> reconcile(P resource, Context<P> context) throws Exception {
-    dependents.reconcile(resource, context);
+    initContextIfNeeded(resource, context);
+    dependents.forEach(dependent -> dependent.reconcile(resource, context));
 
     return metrics().timeControllerExecution(
         new ControllerExecution<>() {
@@ -131,8 +167,12 @@ public class Controller<P extends HasMetadata> implements Reconciler<P>,
 
   @Override
   public List<EventSource> prepareEventSources(EventSourceContext<P> context) {
-    final var dependentSources = dependents.prepareEventSources(context);
-    List<EventSource> sources = new LinkedList<>(dependentSources);
+    List<EventSource> sources = new LinkedList<>();
+    dependents.stream()
+        .filter(dependentResource -> dependentResource instanceof EventSourceProvider)
+        .map(EventSourceProvider.class::cast)
+        .map(provider -> provider.initEventSource(context))
+        .forEach(sources::add);
 
     // add manually defined event sources
     if (reconciler instanceof EventSourceInitializer) {
@@ -267,6 +307,6 @@ public class Controller<P extends HasMetadata> implements Reconciler<P>,
 
   @SuppressWarnings("rawtypes")
   public List<DependentResource> getDependents() {
-    return dependents.getDependents();
+    return dependents;
   }
 }
